+++ docs/zh/guide/intro/history.md
@@ -2,6 +2,7 @@
 
 ## V2.1.6
 - **[特性]**  `@voerkai18n/cli`的`translate`执行时提供错误信息增强
+- **[BUG]** 修复`@voerkai18n/vue2`的`activeLanguage`不会实时渲染更新的问题。
 
 ## V2.1.4